Fix StyleCop warnings in Choise\Parsers -Remove extra black space -Add trailing coma in multi-line initializers -Separate the file ChoiseParser on to   --OptionsOtherMatchParseResult   --OptionsParser   --OptionsParseDataResult
