+++ .NET/Microsoft.Recognizers.Text.Choice/Parsers/OptionsOtherMatchParseResult.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace Microsoft.Recognizers.Text.Choice
+{
+    public class OptionsOtherMatchParseResult
+    {
+        public double Score { get; set; }
+
+        public string Text { get; set; }
+
+        public object Value { get; set; }
+    }
+}
