+++ .NET/Microsoft.Recognizers.Text.Choice/Parsers/OptionsParseDataResult.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace Microsoft.Recognizers.Text.Choice
+{
+    public class OptionsParseDataResult
+    {
+        public double Score { get; set; }
+
+        public IEnumerable<OptionsOtherMatchParseResult> OtherMatches { get; set; }
+    }
+}
